+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.2
+
+* Added example.
+* Fixed README files.
+* Added getCopiedAllTags method to ExifReader.
+* Added deepCopy method to ExifTag.
+
+
 ## 0.0.1
 
 * Initial release.